+++ src/lib/utils/logger.utils.spec.ts
@@ -18,15 +18,15 @@ describe('NGXLoggerUtils', () => {
 
     it('should return custom values if config is provided', () => {
       const config = new LoggerConfig();
-      config.colorScheme = ['#800080', '#008080', '#808080', '#808080', '#FF0000', '#FF0000', '#FF0000'];
+      config.colorScheme = ['violet', 'indigo', 'blue', 'green', 'yellow', 'orange', 'red'];
 
-      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.TRACE, config.colorScheme)).toBe('#800080');
-      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.DEBUG, config.colorScheme)).toBe('#008080');
-      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.INFO, config.colorScheme)).toBe('#808080');
-      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.LOG, config.colorScheme)).toBe('#808080');
-      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.WARN, config.colorScheme)).toBe('#FF0000');
-      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.ERROR, config.colorScheme)).toBe('#FF0000');
-      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.FATAL, config.colorScheme)).toBe('#FF0000');
+      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.TRACE, config.colorScheme)).toBe('violet');
+      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.DEBUG, config.colorScheme)).toBe('indigo');
+      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.INFO, config.colorScheme)).toBe('blue');
+      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.LOG, config.colorScheme)).toBe('green');
+      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.WARN, config.colorScheme)).toBe('yellow');
+      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.ERROR, config.colorScheme)).toBe('orange');
+      expect(NGXLoggerUtils.getColor(NgxLoggerLevel.FATAL, config.colorScheme)).toBe('red');
       expect(NGXLoggerUtils.getColor(NgxLoggerLevel.OFF, config.colorScheme)).toBeUndefined();
     });
   });